Scene structure module
