+++ src/main/python/plotlyst/view/widget/scene/__init__.py
@@ -1,0 +1,1 @@
+from .structure import SceneStructureWidget
